requirements.txt
Pillow==9.5.0
Pillow==9.5.0
PyPDF4==1.27.0
PyPDF4==1.27.0
streamlit==1.24.0
